test app and provider create
